Adding a test to confirm help exists on machine.
Making updates based on feedback from @vors
