+++ src/Markdown.MAML/Renderer/Markdownv2Renderer.cs
@@ -77,101 +77,7 @@ namespace Markdown.MAML.Renderer
                     RenderCleaner.NormalizeQuotesAndDashes(
                         _stringBuilder.ToString())));
         }
-
-        public string AboutMarkdownToString(DocumentNode document)
-        {
-            //ensure that all node types in the about topic are handeled.
-            List<MarkdownNodeType> AcceptableNodeTypes = new List<MarkdownNodeType>();
-            AcceptableNodeTypes.Add(MarkdownNodeType.Heading);
-            AcceptableNodeTypes.Add(MarkdownNodeType.Paragraph);
-            AcceptableNodeTypes.Add(MarkdownNodeType.CodeBlock);
-            if (document.Children.Any(c => (!AcceptableNodeTypes.Contains(c.NodeType)) ))
-            {
-                throw new NotSupportedException("About Topics can only have heading, parapgrah or code block nodes in their Markdown Model.");
-            }
-
-            //processes all nodes in order
-            foreach (var currentNode in document.Children)
-            {
-                switch (currentNode.NodeType)
-                {
-                    case MarkdownNodeType.Paragraph:
-                        ParagraphNode paragraphNode = currentNode as ParagraphNode;
-                        AddAboutParagraph(paragraphNode);
-                        break;
-                    case MarkdownNodeType.Heading:
-                        HeadingNode headingNode = currentNode as HeadingNode;
-                        AddAboutHeading(headingNode, document);
-                        break;
-                    case MarkdownNodeType.CodeBlock:
-                        CodeBlockNode codeblockNode = currentNode as CodeBlockNode;
-                        AddAboutCodeBlock(codeblockNode);
-                        break;
-                }
-            }
-            
-            return RenderCleaner.FullNormalization(_stringBuilder.ToString());
-        }
-
-        private void AddAboutCodeBlock(CodeBlockNode codeblockNode)
-        {
-            string[] lines = codeblockNode.Text.Split(new string[] {"\r\n"}, StringSplitOptions.None);
-            
-            foreach (string line in lines)
-            {
-                _stringBuilder.Append("    ");
-                _stringBuilder.AppendLine(line);
-            }
-
-            _stringBuilder.AppendLine();
-        }
-
-        private void AddAboutHeading(HeadingNode headingNode, DocumentNode document)
-        {
-            if (document.Children.ElementAt(0) as HeadingNode == headingNode)
-            {
-                _stringBuilder.AppendLine("TOPIC");
-            }
-            else if (headingNode.HeadingLevel == 1)
-            {
-                _stringBuilder.AppendLine(headingNode.Text.ToUpper());
-            }
-            else if (headingNode.HeadingLevel == 2 && document.Children.ElementAt(1) == headingNode)
-            {
-                _stringBuilder.AppendLine("    " + headingNode.Text.ToLower() + "\r\n");
-            }
-            else if (headingNode.HeadingLevel == 2)
-            {
-                _stringBuilder.AppendLine(headingNode.Text);
-            }
-            else
-            {
-                _stringBuilder.AppendLine("    " + headingNode.Text.ToUpper());
-            }
-        }
-
-        private void AddAboutParagraph(ParagraphNode paragraphNode)
-        {
-            foreach (string lineContent in paragraphNode.Spans.Select(span => span.Text))
-            {
-                //handles all paragraph lines over 80 characters long and not headers
-                if (lineContent.Length > _maxLineWidth - 4)
-                {
-                    WrapAndAppendLines(lineContent, _stringBuilder);
-                    _stringBuilder.AppendLine();
-                }
-                else if (lineContent == "\n")
-                {
-                    _stringBuilder.AppendLine();
-                }
-                else
-                {
-                    _stringBuilder.AppendLine("    " + lineContent);
-                }
-            }
-            _stringBuilder.AppendLine();
-        }
-
+        
         private void AddYamlHeader(Hashtable yamlHeader)
         {
             _stringBuilder.AppendFormat("---{0}", Environment.NewLine);
@@ -628,72 +534,6 @@ namespace Markdown.MAML.Renderer
                 .Replace(@"`", @"\`")
 
                 ;
-        }
-
-        public string ReplaceMarkdownHeaders(ref int headerLevel, string text)
-        {
-            if (text.Substring(0, 1) == "#")
-            {
-                headerLevel++;
-                text = ReplaceMarkdownHeaders(ref headerLevel, text.Substring(1));
-            }
-
-            return text;
-        }
-
-        public void WrapAndAppendLines(string text, StringBuilder sb)
-        {
-            string[] words = text.Split(' ');
-
-            text = "";
-
-            foreach (string word in words)
-            {
-                if (word.Contains("\r\n"))
-                {
-                    string[] breakLine = word.Split(new string[] { "\r\n" }, StringSplitOptions.RemoveEmptyEntries);
-
-                    foreach (string part in breakLine)
-                    {
-                        if (part == breakLine.Last())
-                        {
-                            text += part + " ";
-                        }
-                        else
-                        {
-                            text += part;
-                            sb.AppendLine("    " + text);
-                            text = "";
-                        }
-                    }
-                }
-                else if (text.Length + word.Length > (this._maxLineWidth - 4))
-                {
-                    if (text.Substring(text.Length - 1) == " ")
-                    {
-                        text = text.Substring(0, text.Length - 1);
-                    }
-                    sb.AppendLine("    " + text);
-
-
-                    text = word + " ";
-                }
-                else
-                {
-                    text += word + " ";
-                }
-                
-            }
-
-            if (text.Length > 0 && text != " ")
-            {
-                if (text.Substring(text.Length - 1) == " ")
-                {
-                    text = text.Substring(0, text.Length - 1);
-                }
-
-                sb.Append("    " + text);
-            }
         }
     }
 }
